// codegen/governance/Governance.types.ts
export type Uint128 = string;
export type Decimal = string;
export type Addr = string;
// ...
export type CosmosMsgForEmpty = {
  bank: BankMsg;
} | {
  custom: Empty;
} | {
  staking: StakingMsg;
} | {
  distribution: DistributionMsg;
} | {
  stargate: {
    type_url: string;
    value: Binary;
    [k: string]: unknown;
  };
} | {
  ibc: IbcMsg;
} | {
  wasm: WasmMsg;
} | {
  gov: GovMsg;
};
// ...
export type Uint64 = string;
// ...
export type ProposalVoteOption = "for" | "against" | "amend" | "remove" | "align";
export interface ProposalMessage {
  msg: CosmosMsgForEmpty;
  order: Uint64;
}
// ...
export type ProposalStatus = "active" | "passed" | "amendment_desired" | "rejected" | "executed" | "expired";
export interface ProposalListResponse {
  proposal_count: Uint64;
  proposal_list: ProposalResponse[];
}
export interface ProposalResponse {
  against_power: Uint128;
  aligned_power: Uint128;
  amendment_power: Uint128;
  delayed_end_block: number;
  description: string;
  end_block: number;
  expiration_block: number;
  for_power: Uint128;
  link?: string | null;
  messages?: ProposalMessage[] | null;
  proposal_id: Uint64;
  removal_power: Uint128;
  start_block: number;
  start_time: number;
  status: ProposalStatus;
  submitter: Addr;
  title: string;
  voting_power: [Addr, Uint128][];
}